【6】对话历史模块 1. 对话历史后端开发（新增对话历史、应用和对话历史关联删除、游标查询、管理员查询 2. 对话记忆（将对话记忆保存到redis中，用caffeine保存AI service（里面包含redis的对话记忆） 3. 目前问题：前端无法展示对话历史

+++ src/main/java/com/wyb/aicodemotherme/common/PageRequest.java
@@ -8,12 +8,12 @@ public class PageRequest {
     /**
      * 当前页号
      */
-    private int pageNum = 1;
+    private Long pageNum = 1L;
 
     /**
      * 页面大小
      */
-    private int pageSize = 10;
+    private Long pageSize = 10L;
 
     /**
      * 排序字段
